Renamed package: io.mats3.util_activemq.MatsLocalVmActiveMq to io.mats3.test.activemq.. + better test DLQ fetcher, handling external standard ActiveMQ
The DLQ fetcher assumed that the MQ was configured with the "specific
DLQ per queue" setup - which it obviously should be as it is superior in
production. However, if you just fire up a freshly downloaded ActiveMQ,
and then directs the tests to use that (read class JavaDoc on
MatsLocalVmActiveMq), some of them fails, as the default standard
ActiveMQ uses a common DLQ for all queues, named "ActiveMQ.DLQ". The
"DLQ fetcher" in MatsTestMqInterface now handles this, by firing up both
a consumer for the standard common one, and for the specific, prefixing
with "DLQ.".