Supporting multiple MQ brokers, also external + Renaming MatsLocalVmActiveMq to MatsTestBroker, and MatsTestMqInterface to MatsTestBrokerInterface ++
* Due to this support for multiple brokers, the MatsLocalVmActiveMq
class, which I've never like the name of, became violently wrongly
named. Thus, changed that to MatsTestBroker. And then also changed the
name of MatsTestMqInterface to MatsTestBrokerInterface, to more directly
correlate those.

* Created a new package for MatsTestBroker, mats-test-broker, instead of
the now deprecated mats-test-activemq (that held MatsLocalVmActiveMq).

* The MatsTestBrokerInterface now fires up three threads listening for
DLQ messages: The original for the "specific DLQ queue"
DLQ.<originalqueuename>, and one for the ActiveMQ default common DLQ
"ActiveMQ.DLQ", and one for the Artemis default common DLQ "DLQ".

* JmsMatsStatics: Had to change name of some message properties that are
tacked on the actual JMS message, e.g. "mats.traceId" - as such "dot
notation" didn't fly with Artemis. It is now named "mats_traceId".
Luckily, these are not really used for anything yet (outside of possibly
manually introspecting the messages on the broker), so this will have no
impact.

* Due to problems with DLQing when going to Artemis, the test
Test_ThrowExceptionsInServiceShouldDlq now accepts different number of
expected deliveries. Read its code comments. Also it logs more.

* Removed the inner implementation in MatsFactory of MatsWrapper called
MatsWrapperDefault - as it really is not used by the Mats API afterall.
It now resides in mats-util.

* Handling a bug in JmsMatsStageProcessor, where a catch expected the
exception to have a cause. In a specific race situation wrt. shutdown
(closing the session right before the processor had entered the
receive()), the javax.jms.IllegalStateException does not have a cause,
and thus threw NPE, which threw the thread, ending up logging on
std.err.

* The MatsTestBroker now has two system parameters it reads,
"mats.test.broker" and "mats.test.brokerurl", which in combination
chooses which JMS Client (ConnectionFactory) to employ, and where to
connect: in-vm (fires up an in-vm ActiveMQ or Artemis broker), localhost
(shortcut url), or a specific URL. The build.gradle comments explains
how to set these props when running the test suite.

* When using external broker, it is important that the tests are very
clean and consume all messages that are emitted, to services or DLQs.
Fixed this in some tests.

* Changed concurreny for testing-specific TestSpringMatsFactoryProvider
from 1 to 2 - this had evidently been forgotten when I changed this for
the JUnit "Rule_Mats" stuff. A bit of concurrency is good to at least
have a chance to catch stuff where you accidentally rely on no
concurrent messages.

* Had to throw in @DirtiesContext on a couple of the mats-spring-test
tests for when running with external broker. The observed situation was
that the MatsTestLatch was the same, correct instance between Spring env
and the instances injected in the actual test class when running with
in-vm, while the instance injected in the test class "stuck" when
running with external. This makes no sense for me now, but
@DirtiesContext sledgehammered it to work in both situations, so forget
it for now.

* For the Two-MatsFactories test (AppMain_TwoMatsFactories), I had to
make a solution when running with external broker: They fail (quite
obviously) if they connect to the single same external broker.
Therefore, there is a "hack" in MatsTestBroker (a special factory
method) that lets you specifically request an in-vm brokere even though
those properties say that you should connect to external broker.

* Corrected the VerifyShutdownOrderUsingFactoryBeanTest to actually
check for the shutdown order..

*