Adding waitForReceiving() to jmsMatsFactory.start() procedure, plus even more granular shutdown procedure.
Found a test where a topic message was sent before the thread had
started receiving, thus failing. Decided to add full
allEndpoints.waitForReceiving() to the start() method.

Also, Test_IsSubscription (which just creates endpoints, and then test
whether it is a subscription or not, thus being VERY fast) often ended
up in "chillWait()" due to bad concurrency where it was closed while it
was trying to create initial JMS objects. Changed the shutdown procedure
to have a new step0 where the _runFlag is set to false, so that this is
done immediately for all stage processors before starting to evaluate
whether it has gotten to receive etc.