Reducing number of letters in the random part of the flowid
It finally occurred to me the timestamp in milliseconds are included in
the flowid. Therefore, to get a collision, we need to produce the
colliding ids within a millisecond.

Playing a bit with the birthday paradox, I find that with 10 random
letters of the 62-entry alphabet (half the length as it was), if we
produce 1 million ids within a single millisecond, there is less chance
of getting a collision than the chance of a given person in the US
getting hit by lightning in a given year. (Given that I understand
anything of this..)

If anyone ever approaches 1 million flowIds per millisecond, I'll
increase the length.