Copied the MatsTraceStringImpl into a MatsTraceFieldImpl<Z>, to be used for both string and binary
The default MatsTraceStringImpl in 'mats-serial' is a generic
field-based implementation of the data structures needed for the
MatsTrace interface, except for the fact that it also decided which
format the "inner" payloads, i.e. DTOs and STOs would be (also String).
Refactored this to instead use generic type Z.

Then made a new class MatsTraceStringImpl in the 'mats-serial-json'
package, simply extending MatsTraceFieldImpl, setting Z to String.

Left the original MatsTraceStringImpl "soft deprecated" in
'mats-serial'.