Fixed bug in Test_MultiLevelStash - required a bit of change.
I have .. "inadvertently" .. introduced even more mutable state: Since I
now set the "StageEnteredTimestamp" on the MatsTrace /upon reception/
(to have the data needed for EndpointEnteredTimestamp calculation), it
is immediately immutable.

The stash functionality relied on using the /incoming/ MatsTrace /in raw
form/, i.e. the actual byte array, before being deserialized, and
notably, therefore /before/ this new setting of EndpointEnteredTimestamp
(long[] matsTrace.eets). This doesn't thus work anymore, and the stash
functionality needs to do an actual serialization of the /now current/
incoming MatsTrace.

This is unfortunate, but I guess I'll have to live with it if I want
this functionality.