Closes #64: Null incoming state: From an initiate, perform a request with replyTo(multistage, null) leads to problems
Fixing bug reported by Ståle Undheim.

If an initiation request is performed to some EndpointX, with
init.replyTo(EndpointA, null), i.e. the replyState being null, EndpointA
was actually invoked with null state. This on first blush seems
intuitive, but it screws over what EndpointA expects: It has declared
some state class S, and expects to be able to set variables on it etc -
it makes no sense that it suddenly is null. Also, Mats internals
(specifically MatsTraceFieldImpl) also expects it to be non-null if the
Endpoint is multi-stage, as it should serialize it onto the stack upon
REQUEST or NEXT. Therefore, a state object cannot ever be null, and in
the specific situation with replyState = null, this must instead create
a new empty state object.