Changing Thread.sleep(100) to more correct endpoint.waitForReceiving(..)
It occurred to me that the fix in the previous commit might be a problem
with other tests too. Evidently, I had realized this problem on the
"SimplePublishSubscribe" test, but used a crude Thread.sleep(100) to
ensure that the endpoint was up. There is now a method specifically made
for this situation, so use that instead.