Github Actions Windows runners can evidently "time-skip" ~100 ms from time to time.
If such a large time skip this happens when running the now infamous
Test_MatsFuturizer_Timeouts while adding the 6 different futures with
different timeouts, then the later added, but with earlier timeout, will
actually have a timeout which is later. Eh. That was easier to think
than to write. Point is that the actual timestamp which should be the
future's timeout is calculated when being added. The shortest timeout
(10ms) is added latest. But, if the time has skipped (i.e. either the
world has stopped, or the clock is skipping, I don't know, but the log
entries jump forward by quite a bit), the actual timeout timestamp of
the shortest timeout will _actually_ be later than the first added
(which has timeout of 100ms). Thus, the order of when the timeouts
happen will be correct wrt. when things was added (the 10ms has a
timeout timestamp that is LATER than the 100ms timeout, so they timeout
in the order 100, 10), but it is wrong compared to the test's
expectation (which obviously is 10, 100).

Increasing difference between the different timeouts, hoping that we'll
avoid the problem. Also moving the addition of the shortest timeout
earlier.