Introducing retrying to get this test to pass on Github Action Windows nodes.
The occasional time-skipping on the Windows nodes of Github Actions are
too severe to handle with even larger intervals of the "scheduled
timeouts", which will affect every interactive command line run (which
have never ever failed). Instead handling with 10x retrying.

This is OK since the timeout functionality is not something meant to be
precisely relied on - it is more of a "garbage collection" of the
MatsFuturizer. Thus, if the test passes even after several attempts, it
is totally fine. The test should catch if the algorithm that orders the
timeouts doesn't work anymore - and the test will still catch that since
one must assume that the failure would be permanent and thus fail all 10
attempts.