Curious CI-error: MatsFuturizer_WhichThread says it faulted, but log indicates correct.
This is the logline from the CI run:

03:35:54.422 [MatsFuturizer completer #1] DEBUG
io.mats3.util.MatsFuturizer - #MATS-UTIL# Completing promise from
[OneSingleMessage]: [io.mats3.util.MatsFuturizer$Promise@7c65d068]
{traceId=traceId}

Clearly indicating that it is the thread "MatsFuturizer completer #1"
that completes this promise.

However, the test failed here:

// The "thenAccept" should have been executed on the MatsFuturizer
thread pool.
Assert.assertTrue(completedOnThreadName[0].startsWith("MatsFuturizer
completer"));

This is certainly curious. The only possibility I see, is the musing I
initially had about this: If some thread has gotten to a point where it
waits on the future, the actual completion happens on that thread, and
not on the thread that invokes future.complete, i.e. that it /hands
over/ the value and lets the waiting thread do the completion.

Sadly, the log line that would have shown this was _below_ the assert,
so it of course would not print. Moving the log line to above assert,
including a failure-message in the assert, and increasing the nap time
in hope to provoke the "waiter has arrived at future" situation.