Better consistency for test wait times.
Got a CI error (on Windows with Java 8): One test waited for 500 ms for
something to happen, which was too short time. It then failed. But then
it happened anyway, and due to shared infrastructure, all the rest of
the tests in the class failed too (getting wrong answers, since it got
the answer that the first test failed to get).

Problem here is that waiting for a short time like 500 ms is
meaningless, as that can fail tests that are affected by infrastructural
problems like e.g. slow CI machines, GC kicking in, and so on.
Therefore, changing over to the default 30 seconds.

Also went through all the waits, changing in particular those tests that
waits for something /not/ to happen. These were using different wait
times throughout. Introduced a constant for this instead.

Also a bit of JavaDocs in the MatsTestLatch wrt. default wait time of 30
seconds for expected occurrences.