Closes #21, closes #44, relevant for #22: Implemented context.goTo(endpointId...)
('goto' is a reserved keyword in Java, thus using context.goTo(..))

In #21, which specifies this feature, this is called context.pass(..). I
had actually first done all the renamings needed for using "pass(..)",
but in the end found goTo to be better. A slight problem was that the
MatsTrace.CallType already had GOTO, but not PASS.

does mention that this would be exactly the same as context.pass(..),
i.e. this feature.

anymore.. Keeping it open for now.