Modifying ActiveMQ instantiation a tad.
Adding setTopicPrefix("DLQ.") for IndividualDeadLetterStrategy. However,
this seemingly doesn't work for ordinary (non-durable) topics, so Mats
won't see subscriptionTerminators(..) ever DLQing. AFAIU.

Also deleting the TimedSubscriptionRecoveryPolicy, and not replacing.
I've never seen any indication that this is a problem, and can't really
envision how it would massively negative effect it - while there is a
penalty of using it, both memory and processing.