Starting impl of #33: context.nextDirect()
The test showing that the bare minimum works, but lots of meta missing,
i.e. logging, metrics, and debugging help are all still relating to the
stage performing the nextDirect, not the nextDirect stage being invoked.

(Not very happy with JmsMatsStageProcessor which has gotten rather
crowded and convoluted, and this new function certainly doesn't make it
better.)