Got a GH Action failure where a certain event happened too fast.
This could possibly have been a memory visibility flaw, but since the
threads are joined just above there, there should be a happens-before
edge in play here. So maybe it was the ">" instead of ">=". Or maybe it
was a "time jump" where the clock was adjusted right then.

Adding a bit of "room" in the test by subtracting a bit of the naptime -
the validity of the test should still hold.