Fixing missing sync bug on ScenarioDecider logic that could happen in test scenarios
A small and innocuous refactor in b100f73fec4 revealed that
ScenarioConnectionFactoryWrapper had missing synchronization when being
invoked in "all-lazy" Remock/MockBean scenarios: A first thread would
come in and lazily start the creation of the "scenario decider"
ConnectionFactory wrapper. However, the created wrapper was immediately
set (even though it had not yet been fully created). Another thread
would come in and find and use the unfinished result - specifically, the
startStopWrapper logic had not yet finished and thus the actual
ConnectionFactory was not set on the resulting wrapper.

Also, minor improvement of system property hacking and code comments
there.