Some adjustments to ActiveMQ creator method.
Including a "features enum", to make it a tad more general. The thinking
is that it should be a nice way to explore the broker and how it works,
and that the interface is clean - a no-args invocation gives you an
in-vm ActiveMQ, and then you can ask for more, e.g. LOCALHOST to become
an actual TCP-reachable server.