For metrics, adding new MDC-key with the name of the dynamically constructed metric MDC-key
We got a problem when trying to find all metrics a given stage had
outputted. The problem is that Elastic have no easy way to actually find
a set of field names using the query language, only field values. Thus,
we add the generated metric key into a static MDC-key, so that it will
be easy to find "this stage emits two timings and one metric, on these
keys".