Adding logging for a problem biting a developer: Missing transient marker on injected field
This bit a developer using MatsClassMapping, with Mats set up using GSON
for serialization of STO/DTO. GSON is eagerly walking the full graph of
the class, and then freaks out when it can't get into e.g.
ThreadPoolExecutor's internals (under Java 17). The solution is to add
'transient' to all injected fields, since they should not be serialized
anyway.