Adding stage completed prop "mats.exec.In.TotalWire.bytes", same as incoming.
Been annoying to not have this property when trying to correlate message
sizes with processing times.

I've multiple times wondered whether to just include all the "incoming"
props on the "stage completed" log line too. Not doing it because of
resulting size, but it is annoying that they're missing. So adding props
when I find actual use for it.