Added a more generic Plugin API, which the Intercept API now relies on.
This adds a MatsPlugin interface. A plugin is installed, queried on, and
removed using the FactoryConfig interface. A MatsFactory may choose to
handle MatsPlugins implementing certain interfaces specially - and the
JmsMatsFactory handles the MatsInitiateInterceptor and
MatsStageInterceptor (as it did before, but now through the generic
plugin solution).

The glaring omission of the Intercept API was that these, effectively,
plugins weren't lifecycled along with the MatsFactory. They had a hard
time being "live", with own threads, own endpoints, and hooks to the
factory's endpoints and initiators. With this new plugin solution, a
plugin doesn't have to rely on the init and stage processing
exclusively, but can instead rely on being started and stopped - and
know which initiators are added, and endpoints added and removed.

This is slightly breaking for implementors of interceptors (of which I
know none external), and for users of the MatsMetricLoggingInterceptor's
ability to hide logs (of which I know one, the 'mats-broker-monitor').
Interceptors needs to install a different way, and the interfaces
holding the static fields denoting "hide logs" are moved.