Cap the default concurrency to 10 (from raw 2 x number_of_cpus)
If you by accident pick up a processor with 16 cores, "or worse", you'll
get very many threads per Stage, with correspondingly many Sessions and
possibly Connections. The default would then be concurrency = 32, which
would lead to 64 threads (32 for normal, and 32 for interactive).

This caps that calculation to 10 (20), which probably is still a pretty
high (probably too high) concurrency for many situations.