Fixes #90: Improve use of Deflate in MatsTrace: Avoid intermediate UTF-8 byte array on way to/from compressed byte array.
When starting with MatsEagerCache, I wanted to make a "fully streaming"
solution, from domain objects, via DTOs, serializing the result
one-by-one and compressing "on the way", finally ending in up in a
compressed byte array.

I was surprised to find that the MatsSerializerJson didn't employ this
"serialize directly to compressed byte array" logic, but instead first
serialized to a UTF-8 byte-array, which then was compressed. The same
when decompressing/deserializing: First decompressed to a UTF-8 byte
array, which then was fed into Jackson for deserialization.

So.. This turned into a little reorganzation. Introducing some more
mats-util: DeflateTools, and FieldBasedJacksonMapper.
The first have new DeflaterInputStreamWithStats and
InflaterInputStreamWithStats, which can handle the streaming approach.
The other is a place to set up the Jackson ObjectMapper, since that will
now be needed both for the mats-serial-json and the cache stuff (and
probably for MatsSocket too) - I like that field-only, lenient, compact
config.

The MatsSerializerJson now uses these tools.