Increasing the thundering-herd delay for CI on GHA.
I believe this is a CI problem, where things are extremely sluggish at
times.

It spuriously failed on an assert at both line 129 and 130 of
'Test_EagerCache_RequestFullUpdate', which is the test where we expect
only 1 update having come through after the startup of cache servers and
clients.

The two clients both sends out a "boot" request, and these are supposed
to be "coalesced" so that just one update is being sent. At line 129, we
check that CacheClient #1 has just gotten 1 update (but it had gotten
2), and on line 130, the same test for CacheClient #2. So, both of these
have failed independently. Interestingly, when CacheClient #2 fails
(having gotten 2 updates instead of expected 1), it means that
immediately before, CacheClient #1 was tested and passed, which is
pretty absurd.

Either this is a kind of race triggered by lagging threads on the
famously slow GHA nodes, or it is a bug. I now increase the "short"
delay up to 3 seconds to give the second request longer time to come and
be coaleasced with the first.