So a bit deeper problem: Incorrect election mechanism.
Reason: Incorrect understanding of JMS spec wrt. ordering of messages on
a topic. There is no global order - the only guarantee one have, is that
multiple messages sent *from a single producer* will arrive to all
consumers in the order they was sent. However, messages sent *from
multiple producers* will not gain any global order; Each consumer may
see the messages from producers interleaved in different orders (again:
From each individual producer, the order will be retained as viewed from
each consumer - but each consumer may see the interleaving of these
ordered streams differently).

So now, back to the original idea before this idea of using the MQ to
get a global order: There's now an leader election mechanism in play,
whereby *the lowest* nodename will be the one that carries the process
to conclusion and sends out the actual update. Luckily, since we already
had a thread for coalescing multiple concurrent requests, we just reused
this: Instead of only running it on the node that "was first" (since we
couldn't rely on that), we run it on all nodes - and after the wait,
only the one that won the election (by having the lowest nodename) would
send the next broadcast message to get the actual update sent (including
the elected winner).

This understanding also explained why sometimes it just got stuck: None
of the servers experienced that they were the first, since the message
from S1 got first to S2 (thus "not sent from me"), and the message from
S2 got first to S1 (the same) - and now both thought that they were not
the handling node.

Actually, the reason for sometimes getting two updates (i.e., coalescing
not working) could also be explained by not ensuring that the servers
actually were listening before starting the clients. So one of the
servers would just see the second request - while the other saw two. The
one that saw two would coalese and randomly send or not, but the one
that just saw one had to assume he was alone, so he always sent. This
change thus also have cacheServer[1|2]._waitForReceiving().