Removed the concept of SRC vs. SENT datatype and the explicit mapper
For many of the applications, the source data will probably be read
directly from a database into a DTO meant for the cache clients. The
mapping function would then be identity.

If read from an in-mem domain model, the mapping function can just as
easily be performed on the consumer.accept(..) invocation.

There was an idea here of being very explicit with the distinction
between "source data" and "data sent over the wire", but in the end I
just felt it cluttered the API.