Added Mock solution, and "linking" of Client to Server.
Both are development/testing solutions.

Mocking is for when the service needs a Cache Client available - it can
e.g. be stuck in the Spring context instead of a proper client.

Linking is for when you develop the server and client in the same
codebase. There is a problem here that both needs to make a
SubscriptionTerminator, which is not possible on a single MatsFactory.
But the server can just forward its received messages to the client for
processing instead of the client making its own SubscriptionTerminator.
Which is what "linking" is.