Add Extension_MatsSpring
This is a usefull tool for testing Spring annotated classes in a Mats
Jupiter test. It allows for registering annotations with the
MatsFactory, and will clear out the added endpoints after each test (in
case of nested test classes).

This extension requires jupiter, and also the mats-test-jupiter package.
This is added as a compileOnly dependency, as it does not make sense to
use this withouth mats-test-jupiter.

I contribute this material in accordance with Storebrand's signed SCA.