Closes #96: MatsEagerCacheServer bug under deploy when unable to produce data
If the server was unable to produce data under startup, it will not be
fully up, as in, not up at all. Only when it has been able to assert
that it can produce data will it start, specifically instantiate and
start the NodeAdvertiser.

The HealthCheck (and HTML GUI) asked for information that was directly
gotten from the NodeAdvertiser. Since this was null when this startup
failure was in effect, dereferencing it produced NPE.

Changed these info fields to return Optionals.

But also ended up adding lots of comments, and small refactors. Slight
adjustment to HTML output wrt. server's seen nodes.