Renamed and moved MatsAnnotatedClass
The MatsAnnotatedClass Rule/Extension, is meant for use in pure Java
tests, not involving Spring, even though it does use Spring internally
to wire up the endpoint, and trigger parsing of the Mats annotated
classes.

Also moved the Rule/Extension to the unit/jupiter test package
respectivly, as this is a more logical place for pure Java test helpers.