Improvement of AbstrqactMatsAnnotatedClass
While starting on tests, I discovered that the API and consistency of
the API was not ideal for the AbstractMatsAnnotatedClass. Adding new
endpoints via a class, was only possible before the test was executed,
and adding actual instances only worked after the test starts.

Changed so that instead one can now add both classes and instances
either when declaring the extension, or within individual tests. We
also
now only keep one application context for each test.