Improved error message on double registration
We know track when we register a mats annotated bean, and detect double
registration of the same class (but not the same endpoint, that is
covered by Mats). This should only occur if you register the same class
on the declaration level and in the method, or in the parent and nested
tests.

Also added test example of Mockito

Added some extra life cycle tests for the Extension_MatsAnnotatedClass,
the same criterias should hold true for jUnit, but it is a bit easier to
test in Jupiter with nested classes. It mainly tests the Abstract class.