Add support in Jupiter to configure via Annotations
This adds the @MatsTest annotation, as well as annotations for creating
test endpoints, and to register annotated classes. This is an
alternative way to configure tests in Jupiter for Mats.

The tests also found that if MatsFactory or MatsFuturizer is a field, it
causes conflict with Spring. Thus AbstractMatsAnnotatedClass was changed
to ignore MatsFactory and MatsFuturizer, as those should not be added to
the Spring context via the fields on the test instance.