Fixed some issues with nested classes
Rewrote the logic in Extension_Mats wrt. nested tests. Test context have
(possibly) parents, and this can be used to resolve the root test
context, and the test context where the Extension_Mats is registered.

This change instead always walks up the context tree to find the
extension, or determine that the extension is not registered at any
level above, and thus the current context is the root Mats context.

Also changed the Extension_MatsRegistration to not trigger when the
annotation is not present, or when there is already an Extension_Mats
present in the test context hierarchy.

Finally, changed AbstractMatsAnnotatedClass to only skip fields that are
an instance of MatsFactory, MatsFuturizer fields do not cause issues.