Ensure that we read fields from all instances
Changed so that in Jupiter, which supports nested tests, we pass in all
the instances under tests, so that we can examine the full hierachy.
This does mean that nested classes will not work the same in jUnit, but
that is not how jUnit 4 is used.

So instead, we skip the enclosing class for injection, but we do use
the full hierarchy as provided by Jupiter for injection. This also works
around Java 21 not always creating the synthetic class.