Support registering endpoints in @MatsTest annotation
This adds support for declaring classes that one wishes to be registered
as endpoints on the MatsTest annotation itself. Thus avoiding the need
to create a separate field for the annotated class, when it is not
needed by the test.

I contribute this material in accordance with the signed SCA.