Closes #102: Increase MAX_NUMBER_OF_METRICS.
For context: This feature is meant to catch "contract violations", and
to protect both the JVM's memory and Prometheus' datastore from
exhaustion. The problem is when users are not following the JavaDoc wrt.
the string InitiatorId when initiating a flow - this is meant to be a
constant, different from each place in the code where a message is
initiated (so that you e.g. can find back to that place when something
goes wrong with a flow). Any dynamic information, e.g. which userId this
flow is initiated on behalf of, shall go in the TraceId. However, it has
happened that this has been misunderstood, and dynamic information like
userId has been put in the InitiatorId string.

This then ends up in a cardinality explosion wrt. metrics: Every userId
gets their own time series.

Thus, there is a "catcher" for this problem: When more than
MAX_NUMBER_OF_METRICS is created, it stops adding, and instead logs a
warning line.

The specific number here was an "initial good and informed guess":

* On the one hand, it is good that it is low, so that we don't flood the
JVM memory and Prometheus with lots of meaningless time series (each
series uses a bit of memory, and storage). Also, so that such a problem
is caught as fast as possible.
* However, on the other hand, it must be large enough so that no normal
situation is caught.

The situation described in the issue proves that this number was on the
low side. A 2.5x increase should give quite a bit of headroom. Making it
configurable makes for both more code, more stuff to think about, thus
more annoyances, for very little actual gain. If an InitiatorId is
actually dynamic, it will eventually be caught no matter how large this
number is.