Completing/fixing the "meta" feature, so that super-MatsSerializers can be made
With the previous solution, you could really only change the outer
MatsTrace implementation, while the inner DTO / STO would be problematic
to change. This since the "meta" was not presented to the
deserializeObject(Object serialized, Class type) method. It now does.

To make it easier to provide the "meta", it now is set on the MatsTrace
itself upon creation or deserialization - since you typically get the
serialized representation from MatsTrace - which then also just as well
could hold the meta it was created/deserialized with.