Closes #103: DLQ in MBM with "7 attempts" even though no recorded delivery attempt
1. Added MDC traceId and a few more so that if this happens, the log
will have a trace that it handled the message (which threw off the hunt
for where the message ever went for customer)

2. Added a new "phase" in the shutdown sequence, before any waiting,
where if the StageProcessor thread is in .receive(), we'll immediately
close the session "from the outside" (i.e. from the main shutdown
thread). Only after all that are in receive() have been closed will we
start waiting for those that aren't there yet. See JmsMatsStageProcessor
stopPhaseX_xx() methods.