Closes #104: Problem with NodeAdvertiser being null on update, and with Healthcheck
The NodeAdvertiser is created lazy - and will thus be null until things
are up. This means that all accesses to it must be guarded by nullcheck.
Evidently, there are some places where it is not - and in certain
situations where stuff doesn't come up in the expected order, it is not
in place when we do get an update.

Change to make it eagerly, upon construction, and then start the active
thread later.

Changed both NodeAdvertiser and PeriodicUpdate classes to be eagerly
created, and started later.