Adding "going away" message when cache server|client shuts down.
Noticed that in customer's setting, where there is a "cluster switch"
concept, both sets of clients are shown in monitor and healthcheck, even
though "the other" set pretty much immediately goes down. That won't
be cleared out until a cache server restart, or 04:00 in the night.

Adding an explicit "going away" advertisement concept, to try to get
them away from the lists.