Adding try-catchall around the coalescing thread.
Customer got into a non-auto-recoverable situation where the JMS Server
was dead when the coalescing thread tried to send the "to phase 2"
message. This then lead to a stale state, where every new message that
came in were "coalesced" into the "existing coalescing thread", but that
was dead! So things escalated, with more and more ensurers trying to get
things going - but these also were just "coalesced".

So, adding a try-catchall that resets the coalescing system if this
happens. Then the ensurer will eventually find that an update never
happened, and kick it off again.