MatsEagerCache: Async bug: If leader dies while producing update, surviving nodes won't reset coalescing.
Resetting of the coalescing system was done by the reception of the
update message sent by the winner of coalescing. Thus, if the leader
that should have sent this message dies (OR something happens with the
network at this point), the coalescing system won't reset, and any
further request for update will forever be coalesced into a state where
it would never exit from.

This is, in hindsight, a rather massive oversight: For certainty wrt.
state transitions on this node: Never trust other nodes (they can go
down), and never trust the network (it might fail).

Now instead doing the reset in "global try-finally" block over the
entire thread code, so that this thread will never exit without having
performed the reset of the coalescing system.