Upgrading Jackson to 3.0.0 (from 2.20.0), which is a major version with new package.
from: com.fasterxml.jackson.databind.ObjectMapper
to:   tools.jackson.databind.ObjectMapper

Some new config, biggest impact is in FieldBasedJacksonMapper which sets
up the Mapper, and that all exceptions now are JacksonException - which
is a RuntimeException (not IOException as before). Otherwise seems ok.

Threw in a few tests, for specific corners and DateTime etc.